+++ lib/hotspots/exit_strategy.rb
@@ -1,41 +1,13 @@
+# compatibility begin
 module Hotspots
-  class ExitStrategy
-    attr_reader :code, :message
+  module OptionBasedExit
+    class Error; end
+    class Safe; end
+    class Noop; end
 
-    def initialize(options)
-      @message = options[:message]
-      @code    = options[:code]
-    end
-
-    def perform
-      puts @message
-      exit @code
-    end
-
-    class Safe
-      attr_reader :code, :message
-
-      def initialize(options)
-        @message = options[:message]
-        @code    = 0
-      end
-
-      def perform
-        puts @message
-        exit @code
-      end
-    end
-
-    class Null
-      attr_reader :code, :message
-
-      def initialize(options = {})
-        @message = ""
-        @code    = nil
-      end
-
-      def perform
-      end
-    end
+    ExitStrategy = Error
+    ExitStrategy::Safe = Safe
+    ExitStrategy::Null = Noop
   end
 end
+# compatibility end